Move AST checks to subfolder
